Use www - see comment for why
The www directory is the directory expected by a Shiny application to locally store the elements that will be rendered in the web browser and are not the output of the scripts. This includes images, HTML, .js files, and so on. This directory must be at the same level as UI.R and server.R, that is, on the application's main folder.